+++ lib/extension/extension.dart
@@ -1,0 +1,8 @@
+library;
+
+export 'extension_double.dart';
+export 'extension_json.dart';
+export 'extension_list.dart';
+export 'extension_map.dart';
+export 'extension_string.dart';
+export 'extension_widget.dart';