+++ lib/flutter_smart.dart
@@ -1,1 +1,5 @@
+library;
+
+export 'extension/extension.dart';
 export 'utils/utils.dart';
+export 'widget/widget.dart';